Apply black

+++ utils/dpbench_datagen/l2_distance/generate_data_random.py
@@ -6,21 +6,23 @@ SEED = 777777
 DTYPE = np.float32
 
 
-#write input data to a file in binary format
+# write input data to a file in binary format
 def __dump_binary__(data_array, file_name):
-    with open(file_name, 'w') as fd:
+    with open(file_name, "w") as fd:
         data_array.tofile(fd)
 
 
-#write input data to a file in text format
+# write input data to a file in text format
 def __dump_text__(data_array, file_name):
-    with open(file_name, 'w') as fd:
-        data_array.tofile(fd, '\n', '%s')
+    with open(file_name, "w") as fd:
+        data_array.tofile(fd, "\n", "%s")
 
 
 def gen_data(nopt, dims, dtype=DTYPE):
     rnd.seed(SEED)
-    return rnd.random((nopt, dims)).astype(dtype), rnd.random((nopt, dims)).astype(dtype)
+    return rnd.random((nopt, dims)).astype(dtype), rnd.random((nopt, dims)).astype(
+        dtype
+    )
 
 
 # call numpy.random.uniform to generate input data and write the input as binary to a file
